+++ docs/2.developers/7.showcases/.user_pairs_fuzzy_join/article.py
@@ -45,7 +45,7 @@ def table_from_kafka(**kwargs):
         receiver=str,
         currency=str,
         amount=int,
-        timestamp=pw.DATE_TIME_UTC,
+        timestamp=pw.DateTimeUtc,
         usd_estimate=int,
     )